Refactored file organization
