+++ Pi App Code/Objects/enums.py
@@ -1,16 +1,5 @@
 from enum import Enum
 
-class GamePiece():
-    def __init__(self, player, location):
-        self.player = player
-        self.location = location
-        self.isKing = False
-
-class Move():
-    def __init__(self, start, end):
-        self.start = start
-        self.end = end
-
 class Player(Enum):
     BLACK = 1
     RED = 2
